Fixed pom dependencies by moving them out of plug dependency section
The main problem is that all dependencies are declared as plugin dependencies, not as regular dependencies. This means, when one of these dependencies e.g. references a different kubernetes-model then the plugin itself, nasty "MethodNotFoundException" occur. The solution is to move it outside of the plugin definition to regular dependencies. Then it works smoothly and also these awful Hibernate validator warnings vanish.